client: modularized Step 4 component for ease of access
- renamed to step 3 for clarity in flow
